requirements.txt
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
feedparser==6.0.11
h11==0.14.0
idna==3.6
numpy==1.26.2
outcome==1.3.0.post0
pandas==2.1.4
pycparser==2.21
PyMuPDF==1.23.7
PyMuPDFb==1.23.7
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
selenium==4.16.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.23.1
trio-websocket==0.11.1
tzdata==2023.3
urllib3==2.1.0
wsproto==1.2.0
git+https://github.com/openai/whisper.git
langdetect
pytube
pillow
pytesseract
